fixes #226
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@6902 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/gui/skins/jdgui/views/info/LinkGrabberInfoPanel.java
@@ -21,8 +21,8 @@ public class LinkGrabberInfoPanel extends InfoPanel {
 
     private static final String JDL_PREFIX = "jd.gui.skins.jdgui.views.info.LinkGrabberInfoPanel.";
     private LinkGrabberController lgi;
-    protected int links;
-    protected int tot;
+    protected long links;
+    protected long tot;
     protected ArrayList<LinkGrabberFilePackage> fps = new ArrayList<LinkGrabberFilePackage>();
 
     public LinkGrabberInfoPanel() {
